change timer trigger to 16bit
- timer compares at approx. 25ms now
- pattern adapted to 16bit to better fine tune timing

+++ app/dabling.h
@@ -173,7 +173,7 @@ void show7Segment(uint8_t left, uint8_t right);
  * \param led definition in leds_config.h
  * \param trigger pattern
  */
-void flashLed(eLED led, uint8_t trigger);
+void flashLed(eLED led, uint16_t trigger);
 
 #endif
